# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Lead Verification Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [firstName, company, phoneNumber, status, campaignName, createdAt]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"

    post:
      operationId: createLead
      tags: [leads]
      summary: Create a single lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/export:
    get:
      operationId: exportLeads
      tags: [leads]
      summary: Export leads as CSV data
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
            default: csv
      responses:
        "200":
          description: Export payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /leads/import:
    post:
      operationId: importLeads
      tags: [leads]
      summary: Bulk import leads from parsed file data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadImportInput"
      responses:
        "201":
          description: Import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadImportResult"
  /leads/requeue:
    post:
      operationId: requeueLeads
      tags: [leads]
      summary: Requeue leads for verification (bulk or by filter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadRequeueInput"
      responses:
        "200":
          description: Requeue summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadRequeueResult"
  /leads/assign-campaign:
    post:
      operationId: assignLeadsCampaign
      tags: [leads]
      summary: Bulk assign leads to a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadAssignCampaignInput"
      responses:
        "200":
          description: Assignment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadAssignCampaignResult"
  /leads/delete-bulk:
    post:
      operationId: deleteLeadsBulk
      tags: [leads]
      summary: Bulk delete leads (by selection or filter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadBulkDeleteInput"
      responses:
        "200":
          description: Delete summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadBulkDeleteResult"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/call:
    post:
      operationId: triggerLeadCall
      tags: [leads]
      summary: Trigger AI verification call for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Lead:
      type: object
      required: [id, firstName, lastName, company, phoneNumber, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
          enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadDetail:
      type: object
      required: [id, firstName, lastName, company, phoneNumber, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        createdAt:
          type: string
        callResult:
          $ref: "#/components/schemas/CallResult"
    CallResult:
      type: object
      properties:
        id:
          type: integer
        callDuration:
          type: ["integer", "null"]
        aiClassification:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        recordingUrl:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        calledAt:
          type: ["string", "null"]
    LeadInput:
      type: object
      required: [firstName, lastName, company, phoneNumber]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phoneNumber:
          type: string
        campaignId:
          type: integer
    LeadImportInput:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadInput"
        campaignId:
          type: integer
    LeadImportResult:
      type: object
      required: [imported, failed, skipped, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
          description: Leads skipped because phone number already exists
        total:
          type: integer
        errors:
          type: array
          items:
            type: string
    LeadRequeueInput:
      type: object
      properties:
        leadIds:
          type: array
          items:
            type: integer
        allFiltered:
          type: boolean
          description: When true, requeue all leads matching the filter fields below
        search:
          type: string
        status:
          type: string
          enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        campaignId:
          type: integer
    LeadRequeueResult:
      type: object
      required: [requeued, skipped, total]
      properties:
        requeued:
          type: integer
        skipped:
          type: integer
          description: Leads skipped (e.g. currently on a call)
        total:
          type: integer
    LeadAssignCampaignInput:
      type: object
      required: [targetCampaignId]
      properties:
        targetCampaignId:
          type: integer
          description: Campaign to assign selected or filtered leads to
        leadIds:
          type: array
          items:
            type: integer
        allFiltered:
          type: boolean
          description: When true, assign all leads matching the filter fields below
        search:
          type: string
        status:
          type: string
          enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        campaignId:
          type: integer
          description: Filter by current campaign
    LeadAssignCampaignResult:
      type: object
      required: [assigned, skipped, total]
      properties:
        assigned:
          type: integer
        skipped:
          type: integer
          description: Leads skipped (e.g. currently on a call)
        total:
          type: integer
    LeadBulkDeleteInput:
      type: object
      properties:
        leadIds:
          type: array
          items:
            type: integer
        allFiltered:
          type: boolean
          description: When true, delete all leads matching the filter fields below (or all leads when no filters)
        search:
          type: string
        status:
          type: string
          enum: [pending, calling, verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        campaignId:
          type: integer
    LeadBulkDeleteResult:
      type: object
      required: [deleted, skipped, total]
      properties:
        deleted:
          type: integer
        skipped:
          type: integer
          description: Leads skipped (e.g. currently on a call)
        total:
          type: integer
    LeadListResponse:
      type: object
      required: [leads, total, page, limit]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ExportResponse:
      type: object
      required: [data, filename]
      properties:
        data:
          type: string
        filename:
          type: string
        count:
          type: integer
...